Replace SQLite dependency with optional use of memcached
Since the SQLite implementation caches to disk and also adding sqlite component requiered recompjlation of python, it was very cumbersome to use. This change removes the dependency on SQlite and uses an in memory cache instead. There are two caches, a local in memory cache and optionally if memcached is available, it will be used as a secondary cache. This secondary cache is used when ansible switches python processes that would result in losing the primary in process local memory cache.